refactor: move `3dMarker` and `3dMarkers` from root to Renderer folder and fix include paths
