+++ source/StdInc.h
@@ -234,8 +234,8 @@
 #include "WaterCreature_c.h"
 #include "WaterCreatureManager_c.h"
 #include "CdStreamInfo.h"
-#include "3dMarker.h"
-#include "3dMarkers.h"
+#include "Renderer/3dMarker.h"
+#include "Renderer/3dMarkers.h"
 #include "Accident.h"
 #include "AccidentManager.h"
 #include "AttractorScanner.h"
